requirements.txt
requests
flask==0.12
flask-assets
#-e git+https://github.com/abourget/gevent-socketio.git#egg=gevent-socketio
#gevent-socketio
gevent-websocket
gevent
pyyaml
redis==2.10.5
msgpack-python
flask-openid
fabric
gunicorn
websocket-client
pympler
soundcloud
spotipy
python-dateutil
jinja2
itsdangerous
werkzeug
click
psycopg2-binary
Cython==0.22.1
pyzmq==14.0
protobuf==2.6.1
